Adjust the handling of the ServerAction clients. Add extra functions to create the correct objects (copying the cookie wrapping) in supabaseServer.ts
Avoid calling a ServerAction function not as a ServerAction.
Replace that call with a direct use of the existing ServerComponent client, which avoids triggering the cookie error.

Remove the 'ServerAction' getIsOrganizer function.
It is unlikely this would ever actually be called as a ServerAction, and was incorrectly grouped here on the basis that it would be called on the server.